requirements.txt
Flask
tensorflow
numpy
Pillow
gunicorn
python-dotenv